nits

+++ pkg/math/oracle/utils.go
@@ -18,7 +18,7 @@ const ScaledDecimals = 40
 // standard number of decimals. We scale the price to the standard number of decimals for ease
 // of comparison.
 func ScaleUpCurrencyPairPrice(decimals int64, price *big.Int) (*big.Int, error) {
-	diff := int64(ScaledDecimals - decimals)
+	diff := ScaledDecimals - decimals
 	if diff < 0 {
 		return nil, fmt.Errorf("cannot scale down price with more decimals than the standard: max=%d, current=%d", ScaledDecimals, decimals)
 	}
@@ -35,7 +35,7 @@ func ScaleUpCurrencyPairPrice(decimals int64, price *big.Int) (*big.Int, error) 
 // NOTE: This function should only be used on prices that have already been scaled to the standard
 // number of decimals. The output of this returns the price to its expected number of decimals.
 func ScaleDownCurrencyPairPrice(decimals int64, price *big.Int) (*big.Int, error) {
-	diff := int64(ScaledDecimals - decimals)
+	diff := ScaledDecimals - decimals
 	if diff < 0 {
 		return nil, fmt.Errorf("cannot scale down price with more decimals than the standard: max=%d, current=%d", ScaledDecimals, decimals)
 	}
